Error handling + missing tests
