+++ common/src/file.rs
@@ -3,14 +3,14 @@ use rmp_serde::Deserializer;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Default, Serialize, Deserialize, PartialEq, Clone)]
-pub struct Database {
+pub struct File {
   pub blocks: Vec<Block>,
 }
 
-impl Database {
-  pub fn from_u8(bytes: &[u8]) -> Result<Database> {
+impl File {
+  pub fn from_u8(bytes: &[u8]) -> Result<File> {
     let mut de = Deserializer::new(bytes);
-    let file: std::result::Result<Database, rmp_serde::decode::Error> =
+    let file: std::result::Result<File, rmp_serde::decode::Error> =
       Deserialize::deserialize(&mut de);
     match file {
       Ok(file) => Ok(file),
