+++ runtime/src/lib.rs
@@ -1,4 +1,4 @@
 mod runtime;
-pub use cuentitos_common::Database;
+pub use cuentitos_common::File;
 pub use runtime::Block;
 pub use runtime::Runtime;
